created a new folder: sidebar/addGroupDialgo and decomposed AddGroupDialog into its components
