+++ client/src/components/sidebar/addGroupDialog/dummyData.tsx
@@ -1,0 +1,13 @@
+import { FriendType } from "./AddGroupDialog";
+
+
+export const friends: FriendType[] = [
+  { name: 'Shaam', zID: 532445 },
+  { name: 'Ray', zID: 523495 },
+  { name: 'hhlu', zID: 584290 },
+  { name: 'Sohum', zID: 523840 },
+  { name: 'Chanel', zID: 542567 },
+  { name: 'Nikki', zID: 524596 },
+  { name: 'Micky', zID: 523948 },
+  { name: 'Jasmine', zID: 540938 },
+];
